requirements.txt
kafka-python
psycopg2-binary
sqlalchemy
pandas
python-dotenv
faker
